+++ examples/oapp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/oapp-example
 
+## 0.6.2
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.6.1
 
 ### Patch Changes